[Feat] #7 - 상품 진열 뷰의 UIPageControl을 구현합니다.
